% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCLtoLowerCS.R
\name{addCLtoLowerCS}
\alias{addCLtoLowerCS}
\title{Add CL data to a lower-level CS table in an RDBESDataObject}
\usage{
addCLtoLowerCS(
  rdbes,
  strataListCS,
  strataListCL,
  combineStrata = T,
  lowerHierarchy = "C",
  CLfields = c("CLoffWeight"),
  verbose = FALSE
)
}
\arguments{
\item{rdbes}{An object of class \code{RDBESDataObject}. This object contains tables from the RDBES
data structure, including a \code{CL} table and \code{CS} data.}

\item{strataListCS}{A named list of filter criteria for subsetting the \code{CS} data in the \code{RDBESDataObject}.
The names of the list should match column names in any of the \code{CS} tables.}

\item{strataListCL}{A named list of filter criteria for subsetting the \code{CL} data in the \code{RDBESDataObject}.
The names of the list should match column names in the \code{CL} table.}

\item{combineStrata}{Logical, if \code{TRUE}, strata in the \code{strataListCS} are combined using a vertical bar (\code{|}).}

\item{lowerHierarchy}{A character string specifying the level of the lower hierarchy table to which the CL data will be added.
Currently, only "C" is supported ie BV data only.}

\item{CLfields}{A character vector of field names from the \code{CL} table that will be summed and added as new columns to the lower-level biological data.}

\item{verbose}{Logical, if \code{TRUE}, the function prints informative text.}
}
\value{
A data.table containing the biological data from the lower hierarchy with added strata information from the \code{CL} table and
the sum of the specified fields from the \code{CL} data.
}
\description{
This function adds data from a \code{CL} table in an \code{RDBESDataObject} to a
BV or FM table. It combines information from the \code{CS} and \code{CL} tables and
calculates aggregate statistics such as the sum of the specified fields in the \code{CL} table.
}
\details{
The function first subsets the biological data in the \code{RDBESDataObject} based on the criteria in \code{strataListCS}. It then retrieves
the corresponding \code{CL} data based on the criteria in \code{strataListCL}, sums the fields specified in \code{CLfields}, and adds them as new columns
to the biological data. If \code{combineStrata} is \code{TRUE}, strata columns from the \code{CS} data are collapsed using a vertical bar (\code{|}). The function
currently supports only biological data at the "C" hierarchy level.
}
\examples{
\dontrun{
strataListCS <- list(LEarea="27.3.d.28.1",
                     LEmetier6 = "OTM_SPF_16-31_0_0",
                     TEstratumName = month.name[1:3],
                     SAspeCodeFAO = "SPR")
strataListCL <- list(CLarea="27.3.d.28.1",
                     CLquar = 1,
                     CLmetier6 = "OTM_SPF_16-31_0_0",
                     CLspecFAO = "SPR")
biolCL <- addCLtoLowerCS(rdbesObject, strataListCS, strataListCL,
                         combineStrata = TRUE,
                         lowerHierarchy = "C",
                         CLfields = c("CLoffWeight"))
}

}
\seealso{
\code{\link{getLowerTableSubsets}}, \code{\link{upperTblData}}
}
