% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyGenerateProbs.R
\name{applyGenerateProbs}
\alias{applyGenerateProbs}
\title{Generate probabilities missing from RDBES Data}
\usage{
applyGenerateProbs(x, probType, overwrite, runInitialProbChecks = TRUE)
}
\arguments{
\item{x}{- RDBES raw object}

\item{probType}{- string. Can be set to "selection" (only selection
probabilities are calculated), "inclusion" (only inclusion probabilities are
calculated) or "both" (both types of probabilities are calculated)}

\item{overwrite}{- if TRUE will overwrite probabilties already existing for
SRSWR and SRSWOR}

\item{runInitialProbChecks}{- if TRUE runs runChecksOnSelectionAndProbs}
}
\value{
a list of all the RDBES data tables with probabilites calculated
}
\description{
Wrapper to generate probabilities. The wrapper calls
runChecksOnSelectionAndProbs which main tests need to be passed before
probabilities can be calculated. The it calls generateProbs for
each sample in each sampling level of the hierarchy.
}
\seealso{
\code{\link{runChecksOnSelectionAndProbs}}
\code{\link{generateProbs}}

examples for now see
https://github.com/ices-eg/WK_RDBES/tree/master/WKRDB-EST2/chairs/Nuno
}
