% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.integer.or.dbl}
\alias{as.integer.or.dbl}
\title{as.integer.or.dbl}
\usage{
\method{as.integer}{or.dbl}(x)
}
\arguments{
\item{x}{vector to be coerced to integers or doubles}
}
\value{
a vector of integers or doubles
}
\description{
This function checks if any value in a vector is above 2e+09, and if so runs
`round(as.double())` on it. If not it uses `as.integer()` instead. This is to
get around the 32-bit limitation in R that integers cannot be larger than
around 2e+09, in which case `as.integer` would return an `NA`.
}
\keyword{internal}
