% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_key_column}
\alias{check_key_column}
\title{Check if a column exists in a data table and has unique values}
\usage{
check_key_column(dt, col)
}
\arguments{
\item{dt}{A data table to check}

\item{col}{A character string specifying the name of the column to check}
}
\value{
nothing if the column exists and has unique values, otherwise an error is thrown
}
\description{
This function checks if a specified column exists in a given data table and has unique values.
If the column does not exist or has non-unique values, an error is thrown.
}
\examples{
check_key_column(H1Example$DE, "DEid")
}
