% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRDBESRawObjects.R
\name{combineRDBESRawObjects}
\alias{combineRDBESRawObjects}
\title{Combine Two RDBES Raw Objects
combines 2 rdbesRawObjects into a single rdbesRawObject by merging individual
tables one by one}
\usage{
combineRDBESRawObjects(rdbesRawObject1, rdbesRawObject2)
}
\arguments{
\item{rdbesRawObject1}{The first object to combine}

\item{rdbesRawObject2}{The second object to combine}
}
\value{
the combination of \code{rdbesRawObject1} and  \code{rdbesRawObject2}
}
\description{
Combine Two RDBES Raw Objects
combines 2 rdbesRawObjects into a single rdbesRawObject by merging individual
tables one by one
}
\examples{
\dontrun{

myH1RawObject <-
    createRDBESRawObject(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19")
myH5RawObject <-
    createRDBESRawObject(rdbesExtractPath = "tests\\\\testthat\\\\h5_v_1_19")

myCombinedRawObject <- combineRDBESRawObjects(rdbesRawObject1=myH1RawObject,
                                             rdbesRawObject2=myH5RawObject)
}
}
\seealso{
\link[data.table]{rbindlist}
}
