% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert.col.names}
\alias{convert.col.names}
\title{convert.col.names}
\usage{
convert.col.names(table, new.names = "R.name")
}
\arguments{
\item{table}{Table to change names of}

\item{new.names}{"field.name" or "R.name"}
}
\value{
a vector of strings of new column names
}
\description{
Converts table column names to field name or R name. For now, not actually
being used anywhere, but will update in future to be more useful.
}
\keyword{internal}
