% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESDataObject.R
\name{createRDBESDataObject}
\alias{createRDBESDataObject}
\title{Create an RDBES Data Object}
\usage{
createRDBESDataObject(
  input = NULL,
  listOfFileNames = NA,
  castToCorrectDataTypes = TRUE
)
}
\arguments{
\item{input}{Strings or \code{list} object. The path to the zip file extracted
from RDBES (or multiple zip files - see details), or path to a folder of
\code{csv} files, or a list object in the current environment containing data
frames of each table. If \code{NULL} an emoty \code{RDBESDataObject} is created.}

\item{listOfFileNames}{\code{list} of Strings, Optional. For csv inputs only, and
only required if the csv file names are \emph{not} the default file names used
by RDBES when downloading data (for instance if you created them yourself).
The names should be a \code{list} of the two-letter code for the relevant table
e.g. \code{list("DE" = "DE.csv", "SD" = "SD.csv", etc.)}.  If not supplied then
it is assumed the files have the default file names used by the RDBES data
download ("Design.csv" etc).}

\item{castToCorrectDataTypes}{Logical. If \code{TRUE} then the function will
attempt to cast the required columns to the correct data type.  If \code{FALSE}
then the column data types will be determined by how the csv files are read
in. Default is \code{TRUE}}
}
\value{
A RDBESDataObject
}
\description{
This function lets you create an RDBES Data object in your current R
environment.
}
\details{
The \code{input} should be either:
\itemize{
\item A \code{zip} file downloaded from RDBES (or multiple zip files if you want to include or overwrite tables, for example CL and CE data)
\item A folder containing \code{csv} files downloaded from RDBES (e.g. the unzipped file), or any set of csv files of the RDBES tables.
\item A \code{list} of data frames in the current environment representing different tables in the hierarchy.
\item A \code{NULL} input will return and emopty RDBES data object
}

\emph{\strong{ZIP file inputs}}
This input should be a path to a zip file downloaded from RDBES. Multiple zip
files can be entered if you want to include additional tables, for example CL
and CE. The main hierarchy input should be the first input. E.g. `input =
c("path/to/H1.zip", "path/to/CL.zip"). If any tables in the first input will
be overwritten by other inputs a warning is given.

\emph{\strong{CSV file inputs}}

\emph{\strong{List of data frames inputs}}

\emph{\strong{NULL inputs}}
}
\examples{
myEmptyRDBESObject <- importRDBESDataCSV(input = NULL)
}
