% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTableOfRDBESIds.r
\name{createTableOfRDBESIds}
\alias{createTableOfRDBESIds}
\title{Create a table of RDBES Ids}
\usage{
createTableOfRDBESIds(x, addSAseqNums = TRUE)
}
\arguments{
\item{x}{RDBESdataObject}

\item{addSAseqNums}{should SAseqNum be included? Default value is TRUE}
}
\value{
data frame of Ids of all tables in sampling hierarchy
}
\description{
examples for now see
https://github.com/ices-eg/WK_RDBES/tree/master/WKRDB-EST2/chairs/Nuno
}
\examples{
\dontrun{

myH1RawObject <-
  importRDBESDataCSV(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19_13")
  
myTableOfIds<- createTableOfRDBESIds(myH1RawObject)
}
}
