% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DefaultFileNames}
\alias{DefaultFileNames}
\title{The default files names when you download data from the RDBES.}
\format{
A list containing the default file names for each RDBES table
}
\source{
\url{https://rdbes.ices.dk/}
}
\usage{
DefaultFileNames
}
\description{
The default files names when you download data from the RDBES.
}
\keyword{datasets}
