% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{designVariables}
\alias{designVariables}
\title{A dataset containing the RDBES "design variable" names}
\format{
A vector containing the short R names of the RDBES design variables
(without any 2 letter table prefixes)
R field name:
\describe{
  \item{designVariables}{The design variable names}
}
}
\source{
\url{https://sboxrdbes.ices.dk}
}
\usage{
designVariables
}
\description{
A dataset containing the RDBES "design variable" names
}
\keyword{datasets}
