% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doDBestimation.R
\name{doDBestimation}
\alias{doDBestimation}
\title{Create design-based point and variance estimates from
RDBES estimation object (rdbesEstimObj)}
\usage{
doDBestimation(
  x = rdbesEstimObj,
  estimateType = "total",
  pointEstimator = "Unbiased",
  varEstimator = "WRonPSUviaPik",
  stage = 0,
  domainOfinterest = NULL
)
}
\arguments{
\item{x}{a data.frame (or data.table) in rdbesEstimObj format
with value of target variable in column targetValue}

\item{estimateType}{a string with type of estimate. As of now only 
"total" is defined}

\item{pointEstimator}{a string with type of variance estimator. As 
of now only "WRonPSUviaPik" is defined}

\item{stage}{a natural number (0,1,..) with sampling stage of 
estimate. 0 corresponds to DE level.}

\item{domainOfinterest}{list ofdomains of interest (e.g., SAarea). As 
of now only NULL (=no domain estimate) is defined}
}
\value{
a list of values for pointEstimate, varEstimate and
estimation options
}
\description{
Create design-based point and variance estimates from
RDBES estimation object (rdbesEstimObj)
}
\examples{
data(shrimps)
doDBestimation (x = shrimps, estimateType = "total", 
pointEstimator = "Unbiased", varEstimator = "WRonPSUviaPsi",stage = 0, 
domainOfinterest = NULL )
}
