% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim.R
\name{estim}
\alias{estim}
\title{Generic function for estimation of population total and variance}
\usage{
estim(y, enk, enkl, method = "SRSWOR", estFunction, varFunction)
}
\arguments{
\item{y}{numeric variable to be estimated}

\item{enk}{expected value of k}

\item{enkl}{expected value of k, given l}

\item{method}{character selection method code e.g SRSWOR}

\item{estFunction}{the function to use to estimate total given
parameters y and enk}

\item{varFunction}{the function to use to estimate variance given
parameters y,enk and enkl}
}
\value{
list of 7 elements including the population mean, total
(and their variance), the algorithm name used and the I order
inclusion probabilities
}
\description{
Generic function for estimation of population total and variance
}
\examples{
estimMC(c(3, 4, 4, 5), c(4, 4, 4, 4), c(8, 8, 8, 8))
}
