% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateHieararhy.R
\name{estimate}
\alias{estimate}
\title{Gather Estimates for Data Table}
\usage{
estimate(tbl, estimRes, target, stratums = NULL, typeOfMeasure = NULL)
}
\arguments{
\item{tbl}{data table to use for estimation}

\item{estimRes}{a list of 2 values. First, named **parentIdCol**, indicates
the column to use for sub-setting the data  and the second is a vector of
values used for sub-setting.}

\item{target}{string, the value to be estimated}

\item{stratums}{named list, vector of stratum names to use at each level.
Defaults to NULL, meaning using all strums.}

\item{typeOfMeasure}{string, for lower hierarchy several measure types can
be used (e.g. age, length, weight). The type of interest can be specified.
Defaults to NULL meaning all will be used.}
}
\value{
a list of estimations at each level
}
\description{
Gets the mean and variance estimates for a single table aggregating all
of the parent children within the table.
}
