% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateHieararhy.R
\name{estimateHieararhy}
\alias{estimateHieararhy}
\title{Generate Multi-Level Estimations}
\usage{
estimateHieararhy(tbls, estimRes, target, ...)
}
\arguments{
\item{tbls}{a named list of data tables that follow the structure of
RDBESrawObject}

\item{estimRes}{a list of 2 values. First, named **parentIdCol**, indicates
the column to use for sub-setting the data  and the second is a vector of
values used for sub-setting.}

\item{target}{string, the value to be estimated}

\item{...}{
  Arguments passed on to \code{\link[=estimate]{estimate}}
  \describe{
    \item{\code{tbl}}{data table to use for estimation}
    \item{\code{stratums}}{named list, vector of stratum names to use at each level.
Defaults to NULL, meaning using all strums.}
    \item{\code{typeOfMeasure}}{string, for lower hierarchy several measure types can
be used (e.g. age, length, weight). The type of interest can be specified.
Defaults to NULL meaning all will be used.}
  }}
}
\value{
a list of values for the estimation result that has all of the
estimation data on uppre tables included
}
\description{
Generates mean and variance estimations for a list of related data tables.
The input data tables is expected to follow the structure of RDBESrawObject.
The mean an variance estimators are calculated for the target variable.
}
\examples{
hd <- generateTestTbls(LETTERS[1:5], selMeth="SRSWOR",propSamp = 0.5)
startFrom <- list(parentIdCol = "Aid", ids=c(1))
estimateHieararhy(tbls=hd, startFrom, target = "Dy")
}
