% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimMC.R
\name{estimMC}
\alias{estimMC}
\title{Multiple Count Estimator for Population Total}
\usage{
estimMC(y, sampled, total, method = "SRSWOR")
}
\arguments{
\item{y}{numeric variable to be estimated}

\item{sampled}{numeric total number of units sampled}

\item{total}{numeric total number of units int the population}

\item{method}{character selection method code e.g SRSWOR}
}
\value{
list of 5 elements including the population mean, total
(and their variance) and the I order inclusion probabilities
}
\description{
Multiple Count Estimator for Population Total
}
\examples{
estimMC(c(3, 4, 4, 5), c(4, 4, 4, 4), c(8, 8, 8, 8))
}
