% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportEstimationResultsToInterCatchFormat.R
\name{exportEstimationResultsToInterCatchFormat}
\alias{exportEstimationResultsToInterCatchFormat}
\title{Export Estimation Results to InterCatch Exchange Format}
\usage{
exportEstimationResultsToInterCatchFormat(dataToExport, verbose = FALSE)
}
\arguments{
\item{dataToExport}{A data frame containing the estimation results -
this should include the output from the doEstimationForAllStrata function
and already have the the InterCatch columns present.}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}
}
\value{
A character vector representing the flattened InterCatch exchange
format. The vector includes all fields from the HI, SI, and SD components,
ordered by their associated keys, and is suitable for writing to an
InterCatch-formatted exchange file.
}
\description{
This function transforms the estimation results into the InterCatch format.
}
