% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fileExt}
\alias{fileExt}
\title{fileExt}
\usage{
fileExt(x)
}
\arguments{
\item{x}{character vector to be split.}
}
\value{
the second element of the character vector.
}
\description{
This function splits a character vector and returns the second element if the
separator is a dot `.`.
}
\keyword{internal}
