% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAndTidyRDBESDataObject.R
\name{filterAndTidyRDBESDataObject}
\alias{filterAndTidyRDBESDataObject}
\title{Filter and an RDBESDataObject and remove any orphan records in an
RDBESDataObject}
\usage{
filterAndTidyRDBESDataObject(
  RDBESDataObjectToFilterAndTidy,
  fieldsToFilter,
  valuesToFilter,
  killOrphans = FALSE,
  verboseOrphans = FALSE,
  verboseBrokenVesselLinks = FALSE
)
}
\arguments{
\item{fieldsToFilter}{A vector of the field names you wish to check}

\item{valuesToFilter}{A vector of the field values you wish to filter for}

\item{killOrphans}{Controls if orphan rows are removed. Default is \code{FALSE}.}

\item{RDBESDataObjectToFilter}{The \code{RDBESDataObject} to filter}
}
\value{
the filtered input object of the same class as
\code{RDBESDataObjectToFilter}
}
\description{
The returned object will include all rows which either: a) do not included
any of the field names in \code{fieldsToFilter}, or b) do include the field names
and have one of the allowed values in \code{valuesToFilter}.
}
\details{
\code{killOrphans} allows you to remove orphaned rows if set to \code{TRUE}. The
default is \code{FALSE}.
}
\examples{
\dontrun{

myH1RawObject <-
  createRDBESDataObject(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19_13")

myFields <- c("VSencrVessCode", "VDlenCat")
myValues <- c("VDcode_1","VL1518","VL2440")

myFilteredObject <- filterRDBESDataObject(myH1RawObject,
  fieldsToFilter = myFields,
  valuesToFilter = myValues,
   killOrphans = TRUE
)
}


}
