% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOrphansByTable.R
\name{findOrphansByTable}
\alias{findOrphansByTable}
\title{Internal function to identify orphan records in a given
RDBESDataObject table}
\usage{
findOrphansByTable(tableToCheck, objectToCheck, foreignKeyIds, verbose = FALSE)
}
\arguments{
\item{tableToCheck}{The two letter code for the table to check}

\item{objectToCheck}{An RDBESDataObject}

\item{foreignKeyIds}{A vetor of the foreign key field names to check}

\item{verbose}{(Optional) If set to TRUE more detailed text will be printed
out by the function.  Default is TRUE.}
}
\value{
A data frame with the primary keys of the table checked, the two
letter table identifier, and their orphan status.
}
\description{
Internal function to identify orphan records in a given
RDBESDataObject table
}
\section{Development review}{

\itemize{
\item AI-assisted: No
\item Human review: rix133
}
}

