% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixSLids.R
\name{fixSLids}
\alias{fixSLids}
\title{Fixes SLid in SL facilitating joins with SS table.
RDBES SL can be seen as a join of two tables - one that identifies the species list in
terms of SLcou\emph{SLinst}SLspeclistName\emph{SLyear}SLcatchFrac and one that specifies
the taxa (SLcommTaxon SLsppCode) in the list. In SS, SLid remits to the 1st taxa in a species list and
not - as it would be expected - to the species list itself. This function fixes this by creating
a new SLtaxaId variable in SL and assigning all taxa in a species to a single SSid.}
\usage{
fixSLids(RDBESDataObject, verbose = FALSE, strict = TRUE)
}
\arguments{
\item{RDBESDataObject}{A valid RDBESDataObject}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}

\item{strict}{(Optional) This function validates its input data - should
the validation be strict? The default is TRUE.}
}
\value{
an RDBESDataObject with SL ids reworked
}
\description{
Fixes SLid in SL facilitating joins with SS table.
RDBES SL can be seen as a join of two tables - one that identifies the species list in
terms of SLcou\emph{SLinst}SLspeclistName\emph{SLyear}SLcatchFrac and one that specifies
the taxa (SLcommTaxon SLsppCode) in the list. In SS, SLid remits to the 1st taxa in a species list and
not - as it would be expected - to the species list itself. This function fixes this by creating
a new SLtaxaId variable in SL and assigning all taxa in a species to a single SSid.
}
\examples{
# To add
}
