% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNAsUsingSL.R
\name{generateNAsUsingSL}
\alias{generateNAsUsingSL}
\title{Generate NAs in samples using Species List information}
\usage{
generateNAsUsingSL(
  RDBESDataObject,
  targetAphiaId,
  overwriteSampled = TRUE,
  validate = TRUE,
  verbose = FALSE,
  strict = TRUE
)
}
\arguments{
\item{RDBESDataObject}{An RDBESDataObject.}

\item{targetAphiaId}{a vector of aphiaId.}

\item{overwriteSampled}{(Optional) should SAtotalWtMes and SAsampWtMes be set to 0 if
spp recorded but absent from SL? The default is TRUE.}

\item{validate}{(Optional) Set to TRUE if you want validation to be carried out. The
default if TRUE.}

\item{verbose}{(Optional) Set to TRUE if you want informative text on validation printed
out, or FALSE if you don't.  The default is FALSE.}

\item{strict}{(Optional) This function can validate its input data - should
the validation be strict? The default is TRUE.}
}
\value{
RDBES data object where SA was complemented with NAs for species not looked for
(sensu in SL)
}
\description{
Generate NAs in samples using Species List information
}
\examples{
# To be added
}
