% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateProbs.r
\name{generateProbs}
\alias{generateProbs}
\title{Generate vector of selection or inclusion probabilities}
\usage{
generateProbs(x, probType)
}
\arguments{
\item{x}{RDBES data object}

\item{probType}{"selection" or "inclusion" for selection and inclusion
probabilities respectively}
}
\value{
A vector or probabilities
}
\description{
Generate vector of selection or inclusion probabilities
}
\details{
When the selection method is SRSWR selection probabilities are
calculated as \eqn{1 / N} and inclusion probabilities as
\eqn{1 - (1 - 1 / N)^n}. When the selection method is SRSWOR selection
probabilities are not currently implemented. Inclusion probabilities are
calculated as \eqn{n/N}. When the selection method is CENSUS both types of
probabilities are set to 1. Probabilities for selection methods UPSWR and
UPSWOR are not calculated (they need to be supplied by the user). The same
happens with regards to non-probabilistic methods
}
\examples{
\dontrun{
generateProbs(x = Pckg_SDAResources_agstrat_H1[["VS"]], probType = ("inclusion"))
# population size
a<-generateProbs(x = Pckg_SDAResources_agstrat_H1[["VS"]], probType = ("inclusion"))
sum(1/a$VSincProb)
# returns error
generateProbs(x = Pckg_SDAResources_agstrat_H1[["VS"]], probType = ("selection"))

}
}
