% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMissingSSRows.R
\name{generateSSRows}
\alias{generateSSRows}
\title{Private function to generate SS rows}
\usage{
generateSSRows(FOids, speciesListName, catchFra)
}
\arguments{
\item{FOids}{Vector of FOids}

\item{speciesListName}{Name of the species list}

\item{catchFra}{The catch fraction to create}
}
\value{
SS data frame
}
\description{
Private function to generate SS rows
}
