% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTestTables.R
\name{generateTestTbls}
\alias{generateTestTbls}
\title{Generate a List of Related Data Tables}
\usage{
generateTestTbls(tblNames, prevTbls = list(), ...)
}
\arguments{
\item{tblNames}{character vector of table names to be created}

\item{prevTbls}{list of data.tables upstream of the generated table.
Defaults to empty list}

\item{...}{
  Arguments passed on to \code{\link[=makeTbl]{makeTbl}}
  \describe{
    \item{\code{rows}}{numeric number of rows per parent record. Defaults to 4.}
    \item{\code{propSamp}}{numeric proportion of how many of total are
sampled. This is ignored for "CENSUS". Defaults to 0.5}
    \item{\code{selMeth}}{character selection method used. Defaults to "CENSUS".
Others like SRSWR or SRSSWOR can be used as well}
    \item{\code{stratums}}{character vector of the stratum names to be created.
Defaults to c("U"), meaning not stratified.}
    \item{\code{mean}}{numeric the expected mean of the target variable.
The variable is created using \code{\link[stats]{rnorm}} and saved under
 column ending with "y". Defaults to 5.}
  }}
}
\value{
a list of named data.table's
}
\description{
Generates a named list of data tables that follow the structure of
RDBESrawObject. The tables only have columns required for testing
The generate tables
}
\examples{
generateTestTbls(c("A", "B", "C"), seletionMethod = "SRSWOR")
generateTestTbls(LETTERS[1:5]) # makes 5 tables with method CENSUS
}
