% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLinkedDataFromLevel.R
\name{getLinkedDataFromLevel}
\alias{getLinkedDataFromLevel}
\title{Retrieve Linked Data Between RDBES Tables at a Specified Level}
\usage{
getLinkedDataFromLevel(field, values, rdbesTables, level, verbose = FALSE)
}
\arguments{
\item{field}{A character string specifying the field name from which to retrieve linked data. The first two characters of this field indicate the source table.}

\item{values}{A vector of values corresponding to the specified \code{field} for which linked data is to be retrieved.}

\item{rdbesTables}{An \code{RDBESDataObject} containing the relevant RDBES tables. This object should include all tables that may be linked based on the provided \code{field} and \code{level}.}

\item{level}{A character string specifying the target table level from which to retrieve linked data. This must be one of the names within the \code{rdbesTables} object.}

\item{verbose}{Logical flag indicating whether to print detailed information about the data retrieval process. Default is \code{FALSE}.}
}
\value{
The subset of the table at the specified \code{level}.
}
\description{
The \code{getLinkedDataFromLevel} function facilitates the retrieval of linked data between different levels of RDBES tables. Depending on the relative positions of the source and target tables within the \code{RDBESDataObject}, the function determines whether to traverse "up" or "down" the data hierarchy to obtain the desired linked data.
}
\examples{
\dontrun{
# Example 1: Going up in the table hierarchy to retrieve data from the DE table
# Retrieve data from the DE level based on BVid from the BV table
# This returns 1 row from the DE table
getLinkedDataFromLevel("BVid", c(1), H8ExampleEE1, "DE", TRUE)

# Example 2: Going down in the table hierarchy to retrieve data from the SA table
# Retrieve data from the SA level based on DEid from the DE table
# This returns 15 rows from the SA table
getLinkedDataFromLevel("DEid", c(1), H8ExampleEE1, "SA", TRUE)

# Example 3: Going up in the table hierarchy to see the Vessel that caught a specific fish
# Retrieve data from the VS level based on BVfishId from the BV table
getLinkedDataFromLevel("BVfishId", c("410472143", "410472144"), H8ExampleEE1, "VS", TRUE)
}
}
