% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLowerTableSubsets.R
\name{getLowerTableSubsets}
\alias{getLowerTableSubsets}
\title{Get Lower Table from Several Upper Table Fields}
\usage{
getLowerTableSubsets(
  subsets,
  tblName,
  rdbesTables,
  combineStrata = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{subsets}{A named list of vectors. Each vector contains values for a specific upper table field.}

\item{tblName}{A character string specifying the name of the target lower level table.}

\item{rdbesTables}{A RDBESData object containing the tables.}

\item{combineStrata}{A logical value indicating whether to include the strata information in the result.
If \code{TRUE}, and if any strata has more than one value, those values are collapsed into a single string
and appended to the result with a warning.}

\item{verbose}{A logical value indicating whether to print informative text.}
}
\value{
A unique data frame containing the rows of the target lower level table that are associated with
the given values of the upper table field in each subset. If \code{combineStrata = TRUE}, the result will also include
a column for each subset with the corresponding collapsed values.
}
\description{
This function takes a list of subsets, a target lower level table name, and a list of tables.
It returns a unique data frame containing the rows of the target lower level table that are associated with
the given values of the upper table field in each subset. The function can also add the subset values to
the result for reference.
}
\details{
The function recursively intersects the rows of the target lower level table that match
the values from each subset in the upper tables. It then ensures that only unique rows are returned,
based on the ID column of the target table.
}
