% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMissingSSRows.R
\name{getMissingSSCatchFraction}
\alias{getMissingSSCatchFraction}
\title{Private function to find which FO rows are not matching SS}
\usage{
getMissingSSCatchFraction(FOdata, SSdata, catchFra, verbose)
}
\arguments{
\item{FOdata}{The FOdata}

\item{SSdata}{The SSdata}

\item{catchFra}{The catchfra}

\item{verbose}{verbose or not?}
}
\value{
Vector of FOids that aren't matching SS rows
}
\description{
Private function to find which FO rows are not matching SS
}
