% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESEstObject.R
\name{getSubSampleLevel}
\alias{getSubSampleLevel}
\title{Private function to get sub-sample level and top-level SAid for SA data}
\usage{
getSubSampleLevel(SAdata, SAidToCheck, subSampleLevel = 1)
}
\arguments{
\item{SAdata}{The SA data to check}

\item{SAidToCheck}{The SAid to check}

\item{subSampleLevel}{The currrent level of sampling}
}
\description{
Private function to get sub-sample level and top-level SAid for SA data
}
