% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTablesInRDBESHierarchy.R
\name{getTablesInRDBESHierarchy}
\alias{getTablesInRDBESHierarchy}
\title{Returns the tables for a given hierarchy}
\usage{
getTablesInRDBESHierarchy(
  hierarchy,
  includeOptTables = TRUE,
  includeLowHierTables = TRUE,
  includeTablesNotInSampHier = TRUE
)
}
\arguments{
\item{hierarchy}{Integer value between 1 and 13 inclusive}

\item{includeOptTables}{Include any optional tables?  Default value is
TRUE}

\item{includeLowHierTables}{Include the lower hierarchy tables?
Default value is TRUE}

\item{includeTablesNotInSampHier}{Include tables that aren't
sampling units in that hierarcy?  Default value is TRUE}
}
\value{
A vector containing the 2-letter names of the tables in the
requested hierarchy
}
\description{
Returns the tables for a given hierarchy
}
\examples{
getTablesInRDBESHierarchy(5)
}
