% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H1Example}
\alias{H1Example}
\title{A dataset containing test RDBES data for H1 in the RDBESDataObject structure}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data table}
\item{SD}{the Sampling Details data table}
\item{VS}{the Vessel Selection data table}
\item{FT}{the Fishing Trip data table}
\item{FO}{the Fishing Operation data table}
\item{SS}{the Species Selection data table}
\item{SA}{the Sample data table}
\item{FM}{the Frequency Measure data table}
\item{BV}{the Biological Variable data table}
\item{VD}{the Vessel Details data table}
\item{SL}{the Species List data table}
}
}
\usage{
H1Example
}
\description{
A dataset containing test RDBES data for H1 in the RDBESDataObject structure
}
\keyword{datasets}
