% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H5Example}
\alias{H5Example}
\title{A dataset containing test RDBES data for H5 in the RDBESDataObject structure}
\format{
A list containing entries required for H5 RDBES data:
\describe{
\item{DE}{the Design data table}
\item{SD}{the Sampling Details data table}
\item{FT}{the Fishing Trip data table}
\item{OS}{the Onshore Event data table}
\item{LE}{the Landing Event data table}
\item{SS}{the Species Selection data table}
\item{SA}{the Sample data table}
\item{FM}{the Frequency Measure data table}
\item{BV}{the Biological Variable data table}
\item{VD}{the Vessel Details data table}
\item{SL}{the Species List data table}
}
}
\usage{
H5Example
}
\description{
A dataset containing test RDBES data for H5 in the RDBESDataObject structure
}
\keyword{datasets}
