% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icesSpecWoRMS}
\alias{icesSpecWoRMS}
\title{A dataset containing a copy of the icesSpecWoRMS code list. The latest
code list data can be downloaded from https://vocab.ices.dk/}
\format{
A data frame
\describe{
  \item{GUID}{Globally unique identifier assigned by ICES}
  \item{Key}{AphiaID}
  \item{Description}{Scientific name}
  \item{LongDescription}{Ignore}
  \item{Modified}{Date when the code was last updated}
  \item{Deprecated}{IS this still a valid code.  If FALSE the code is
  no longer valid within ICES.}
  \item{DateDownloaded}{E.g. "2023-10-18" }
  ...
}
}
\source{
\url{https://vocab.ices.dk/}
}
\usage{
icesSpecWoRMS
}
\description{
A dataset containing a copy of the icesSpecWoRMS code list. The latest
code list data can be downloaded from https://vocab.ices.dk/
}
\keyword{datasets}
