% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRDBESDataZIP.R
\name{importRDBESDataZIP}
\alias{importRDBESDataZIP}
\title{Import RDBES Downloaded Data}
\usage{
importRDBESDataZIP(filenames, castToCorrectDataTypes = TRUE)
}
\arguments{
\item{filenames}{\itemize{
\item path to zipfile. Multiple paths can be entered if, for
instance, you want to add CL and CE tables to the same object.
}}

\item{castToCorrectDataTypes}{Logical. If \code{TRUE} then the function will
attempt to cast the required columns to the correct data type.  If \code{FALSE}
then the column data types will be determined by how the csv files are read
in. Default is \code{TRUE}.}
}
\value{
a list of all the RDBES data tables The table that are not in input
data are NULL
}
\description{
An internal function used by \code{createRDBESDataObject}. Reads the .zip files
downloadable from the ICES RDBES web page. The function accepts a filepaths
to zip files, unzips and import tables as an object of class
\code{RDBESDataObject}.
}
\examples{
rdbesExtractPath <- "./tests/testthat/h7_v_1_19_18/2022_FPN_FPE_H7.zip"
obj <- importRDBESDataZIP(rdbesExtractPath)
}
\keyword{internal}
