% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.zip}
\alias{is.zip}
\title{is.zip}
\usage{
is.zip(x)
}
\arguments{
\item{x}{character vector with the file extension}
}
\value{
logical value
}
\description{
This function evaluates if the file extension is .zip.
}
\keyword{internal}
