% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowerTblData.R
\name{lowerTblData}
\alias{lowerTblData}
\title{Get lower table data from upper table id}
\usage{
lowerTblData(field, values, tbls, level, verbose = FALSE)
}
\arguments{
\item{field}{A character string specifying the name of the upper table field.}

\item{values}{A vector of values for the upper table field.}

\item{tbls}{A named list of data frames representing the tables.}

\item{level}{A character string specifying the name of the target lower level table.}

\item{verbose}{A logical value indicating whether to print intermediate levels during recursion.}
}
\value{
A data frame containing the rows of the target lower level table that are associated with
the given values of the upper table field.
}
\description{
This function takes an upper table field name, values for that field, a list of tables,
and a target lower level table name. It returns the rows of the target lower level table
that are associated with the given values of the upper table field.
NB! running on the RDBESDataObject will work properly if it is sorted by default the tables are not
in the correct order for a spesific hierarchy.
}
\examples{
# it is important to run these functions only on sorted RDBESDataObject
lowerTblData("TEid", c(4), sort(H8ExampleEE1), "SA", T)

DE <- data.table::data.table(DEid = c(1, 2, 3, 4), SDid = c(1, 2, 3, 4))
SD <- data.table::data.table(SDid = c(1, 2, 3, 4), TEid = c(1, 2, 3, 4))
TE <- data.table::data.table(SDid = c(1, 2, 3, 4), TEid = c(1, 2, 3, 4))
VS <- data.table::data.table(TEid = c(1, 2, 3, 4), VSid = c(1, 2, 3, 4))
LE <- data.table::data.table(VSid = 1:5, LEid = 1:5, value = c(10, 20, 3, 4, 6))
tblsSprat <- list( DE = DE ,SD = SD, TE = TE, VS = VS, LE = LE )

lowerTblData("TEid", c(4), tblsSprat, "LE", T)
}
