% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MadeUpData_for_SL_SA_tests_v1}
\alias{MadeUpData_for_SL_SA_tests_v1}
\title{A made-up dataset for testing manipulations of SA based on SL}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data table}
\item{SD}{the Sampling Details data table}
\item{VS}{the Vessel Selection data table}
\item{FT}{the Fishing Trip data table}
\item{FO}{the Fishing Operation data table}
\item{SS}{the Species Selection data table.
Contains one row with keys equal to the SL keys}
\item{SA}{the Sample data table. Contains one species}
\item{FM}{the Frequency Measure data table}
\item{BV}{the Biological Variable data table}
\item{VD}{the Vessel Details data table}
\item{SL}{the Species List data table.
Contains two species for the same SLcountry, SLinstitute,
SLspeciesListName, SLyear, SLcatchFraction, SLcommercialTaxon,
SLspeciesCode & SLcommercialTaxon == SLspeciesCode}
}
}
\usage{
MadeUpData_for_SL_SA_tests_v1
}
\description{
This data set is created for testing the idea of manipulating Sample data (SA)
based on Species List (SL). It represents the simplest case for testing this
idea. The data set contains two species in SL for the same SLcountry,
SLinstitute, SLspeciesListName, SLyear, SLcatchFraction, SLcommercialTaxon,
SLspeciesCode & SLcommercialTaxon == SLspeciesCode. There is one species in
SA - one row in SS with keys equal to the SL keys.
}
\keyword{datasets}
