% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MadeUpData_for_SL_SA_tests_v1_WGRDBES_EST_TEST_3}
\alias{MadeUpData_for_SL_SA_tests_v1_WGRDBES_EST_TEST_3}
\title{A dataset containing test RDBES data for H1 in the RDBESDataObject structure}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data frame}
\item{SD}{the Sampling Details data frame}
\item{VS}{the Vessel Selection data frame}
\item{FT}{the Fishing Trip data frame}
\item{FO}{the Fishing Operation data frame}
\item{SS}{the Species Selection data frame}
\item{SA}{the Sample data frame}
\item{FM}{the Frequency Measure data frame}
\item{BV}{the Biological Variable data frame}
\item{VD}{the Vessel Details data frame}
\item{SL}{the Species List data frame}
...
}
}
\usage{
MadeUpData_for_SL_SA_tests_v1_WGRDBES_EST_TEST_3
}
\description{
A dataset containing test RDBES data for H1 in the RDBESDataObject structure
}
\keyword{datasets}
