% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeDT}
\alias{makeDT}
\title{makeDT}
\usage{
makeDT(x)
}
\arguments{
\item{x}{A data.frame, or \code{NULL}.}
}
\value{
A data.table, or \code{NULL} if \code{x} is \code{NULL}.
}
\description{
Converts one element of a list of data.frames into a data.table, leaving
\code{NULL} elements as \code{NULL}. Used for each RDBES table in
\code{importRDBESDataDFS()}.
}
\keyword{internal}
