% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateTestTables.R
\name{makeTbl}
\alias{makeTbl}
\title{Generate a Data Table}
\usage{
makeTbl(
  tblName,
  prevTbls = list(),
  rows = 4,
  propSamp = 0.5,
  selMeth = "CENSUS",
  stratums = c("U"),
  mean = 5
)
}
\arguments{
\item{prevTbls}{list of data.tables upstream of the generated table.
Defaults to empty list}

\item{rows}{numeric number of rows per parent record. Defaults to 4.}

\item{propSamp}{numeric proportion of how many of total are
sampled. This is ignored for "CENSUS". Defaults to 0.5}

\item{selMeth}{character selection method used. Defaults to "CENSUS".
Others like SRSWR or SRSSWOR can be used as well}

\item{stratums}{character vector of the stratum names to be created.
Defaults to c("U"), meaning not stratified.}

\item{mean}{numeric the expected mean of the target variable.
The variable is created using \code{\link[stats]{rnorm}} and saved under
 column ending with "y". Defaults to 5.}
}
\value{
a data.table
}
\description{
Generate a Data Table
}
