% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pckg_SDAResources_agsrs_H1}
\alias{Pckg_SDAResources_agsrs_H1}
\title{A RDBESDataObject converted from package SDAResources dataset agsrs}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data table. Contains dummy values with exception of
Design-Variables in VS that match the book}
\item{SD}{the Sampling Details data table. Contains dummy values}
\item{VS}{the Vessel Selection data table. Contains core information of
data(agsrs), VSnumberSampled and VSnumberTotal set according to agsrs
and book pop values, VSunitName is set to a combination of original
agsrs$county, agsrs$state, agsrs$region and row numbers}
\item{FT}{the Fishing Trip data table. Contains dummy values}
\item{FO}{the Fishing Operation data table. Contains dummy values}
\item{SS}{the Species Selection data table. Contains dummy values}
\item{SA}{the Sample data table. Contains the variable measured
agsrs$acres92 in SAtotalWeightMeasured, SAsampleWeightMeasured and
SAconversionFactorMeasLive set to 1, and the domain information,
coded in SAarea}
\item{FM}{the Frequency Measure data table. Not provided}
\item{BV}{the Biological Variable data table. Not provided}
\item{VD}{the Vessel Details data table. Subset to the essential rows}
\item{SL}{the Species List data table. Subset to the essential rows}
}
}
\source{
\url{https://CRAN.R-project.org/package=SDAResources}
}
\usage{
Pckg_SDAResources_agsrs_H1
}
\description{
This data set is derived from the data(agsrs) used in Lohr examples 2.6, 2.7
and 2.11 of SDA book. Information required for example 4.8
(domain estimation) is also added to SA (farmcat <=> SAarea).
VSnumberSampled and VSnumberTotal set according to agsrs and book pop
values. VSunitName is set to a combination of original agsrs$county,
agsrs$state, agsrs$region and row numbers. Table SA contains the variable
measured agsrs$acres92 in SAtotalWeightMeasured, SAsampleWeightMeasured and
SAconversionFactorMeasLive set to 1. Table SA also contains the domain
information, coded in SAarea. Table DE, SD, FT and FO are for the most
dummy tables inserted to meet RDBES model requirements to be aggregated
during estimation tests. Values of mandatory fields have dummy values with
exception of Design-Variables in VS that match the book. BV, FM, CL, and CE
are not provided. SL and VD are subset to the essential rows.
}
\keyword{datasets}
