% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pckg_SDAResources_agstrat_H1_WGRDBES_EST_TEST_1}
\alias{Pckg_SDAResources_agstrat_H1_WGRDBES_EST_TEST_1}
\title{A RDBESDataObject converted from package SDAResources dataset agstrat}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data table. Contains dummy values with exception of
selectionMethod that is set to CENSUS}
\item{SD}{the Sampling Details data table. Contains dummy values}
\item{VS}{the Vessel Selection data table. Contains core information of
data(agstrat), VSstratumName set to agstrat$region, and VSnumberSampled
and VSnumberTotal set according to agstrat, VSunitName is set to a
combination of original agstrat$county, agstrat$state, agstrat$region and
agstrat$agstrat row numbers}
\item{FT}{the Fishing Trip data table. Contains dummy values}
\item{FO}{the Fishing Operation data table. Contains dummy values}
\item{SS}{the Species Selection data table. Contains dummy values}
\item{SA}{the Sample data table. Contains the variable measured
agstrat$acres92 in SAtotalWeightMeasured, SAsampleWeightMeasured and
SAconversionFactorMeasLive set to 1}
\item{FM}{the Frequency Measure data table. Not provided}
\item{BV}{the Biological Variable data table. Not provided}
\item{VD}{the Vessel Details data table. Subset to the essential rows}
\item{SL}{the Species List data table. Subset to the essential rows}
}
}
\source{
\url{https://CRAN.R-project.org/package=SDAResources}
}
\usage{
Pckg_SDAResources_agstrat_H1_WGRDBES_EST_TEST_1
}
\description{
This data set is derived from the data(agstrat) used in Lohr examples 3.2
and 3.6. Table VS is stratified with VSstratumName set to agstrat$region,
and VSnumberSampled and VSnumberTotal set according to agstrat.
VSunitName is set to a combination of original agstrat$county,
agstrat$state, agstrat$region and agstrat$agstrat row numbers.
Table SA contains the variable measured agstrat$acres92 in
SAtotalWeightMeasured, SAsampleWeightMeasured and SAconversionFactorMeasLive
set to 1. Table DE, SD, FT and FO are for the most dummy tables inserted to
meet RDBES model requirements to be aggregated during estimation tests.
Values of mandatory fields have dummy values taken from an onboard programme,
with exception of selectionMethod that is set to CENSUS. BV, FM, CL, and CE
are not provided. SL and VD are subset to the essential rows.
}
\keyword{datasets}
