% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pckg_SDAResources_algebra_H1}
\alias{Pckg_SDAResources_algebra_H1}
\title{A RDBESDataObject converted from package SDAResources dataset algebra}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data table. Contains 1 DE row with
DEstratumName == "Pckg_SDAResources_algebra_H1"}
\item{SD}{the Sampling Details data table. Contains 1 child SD row}
\item{VS}{the Vessel Selection data table. Contains 299 child rows
(the 299 students observed), each associated to its cluster (class),
VSnumberTotalClusters is 187, VSnumberSampledClusters is 12,
VSnumberTotal is Missing}
\item{FT}{the Fishing Trip data table. Just 1:1 links to the final
data (in SA)}
\item{FO}{the Fishing Operation data table. Just 1:1 links to the final
data (in SA)}
\item{SS}{the Species Selection data table. Just 1:1 links to the final
data (in SA)}
\item{SA}{the Sample data table. Each score is a SAsampleWeightMeasured}
\item{FM}{the Frequency Measure data table}
\item{BV}{the Biological Variable data table}
\item{VD}{the Vessel Details data table}
\item{SL}{the Species List data table}
}
}
\source{
\url{https://CRAN.R-project.org/package=SDAResources}
}
\usage{
Pckg_SDAResources_algebra_H1
}
\description{
This data set is derived from a fictional data for an SRS of 12 algebra
classes in a city, from a population of 187 classes. The design is 1-stage
cluster sampling with clusters of unequal sizes. Clusters are classes (class).
Clusters (psu) are unequal sized (Mi). In each cluster, all students are
selected (ssus, nrows). The total number of psus is known (187). The target
variable is score.
}
\keyword{datasets}
