% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pckg_SDAResources_coots_multistage_H1}
\alias{Pckg_SDAResources_coots_multistage_H1}
\title{A Multi-stage RDBESDataObject from package SDAResources dataset coots}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data table. Contains 1 DE row with
DEstratumName == "Pckg_SDAResources_coots_H1"}
\item{SD}{the Sampling Details data table. Contains 1 child SD row}
\item{VS}{the Vessel Selection data table. Contains 184 child rows
(the 184 clutches/psus observed), each associated to its cluster (clutch),
VSnumberTotal is not known, VSnumberSampled is 184}
\item{FT}{the Fishing Trip data table. Contains 368 child rows
(the 368 eggs/ssus measured), each associated to its vessel (clutch),
FTnumbersampled is 2, FTnumberTotal is csize}
\item{FO}{the Fishing Operation data table. Just 1:1 links to the final
data (in SA)}
\item{SS}{the Species Selection data table. Just 1:1 links to the final
data (in SA)}
\item{SA}{the Sample data table. Each volume is a SAsampleWeightMeasured.
ATT volumes are *100000000 to meet type requirement (integer)}
\item{FM}{the Frequency Measure data table}
\item{BV}{the Biological Variable data table}
\item{VD}{the Vessel Details data table}
\item{SL}{the Species List data table}
\item{IS}{the Individual Species table}
}
}
\source{
\url{https://CRAN.R-project.org/package=SDAResources}
}
\usage{
Pckg_SDAResources_coots_multistage_H1
}
\description{
This data set is derived from the data(coots). The design is 2-stage cluster
sampling with clusters of unequal sizes and Npsu not known. Clusters are
clutches of eggs (nests) with at least 2 eggs. In each cluster, the volume
of two eggs is measured. Clusters (psu) are unequal sized. In each cluster,
2 eggs are selected (ssus) and measured. The total number of psus is not
known (a drawback in this example). It is assumed very large (fpc negligible).
The target variable is volume.
}
\keyword{datasets}
