% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pckg_SDAResources_gpa_H1}
\alias{Pckg_SDAResources_gpa_H1}
\title{A RDBESDataObject converted from package SDAResources dataset gpa}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data table. Contains 1 DE row with
DEstratumName == "Pckg_SDAResources_gpa_H1"}
\item{SD}{the Sampling Details data table. Contains 1 child SD row}
\item{VS}{the Vessel Selection data table. Contains 20 child rows
(the 20 observations), each associated to its cluster (suite),
VSnumberTotalClusters is 100, VSnumberTotal is 4 because all elements in
cluster are sampled}
\item{FT}{the Fishing Trip data table. Just 1:1 links to the final data
(in SA)}
\item{FO}{the Fishing Operation data table. Just 1:1 links to the final
data (in SA)}
\item{SS}{the Species Selection data table. Just 1:1 links to the final
data (in SA)}
\item{SA}{the Sample data table. Each gpa score is a
SAsampleWeightMeasured. ATT gpa scores are *100 to meet type requirement
(integer)}
\item{FM}{the Frequency Measure data table}
\item{BV}{the Biological Variable data table}
\item{VD}{the Vessel Details data table}
\item{SL}{the Species List data table}
}
}
\source{
\url{https://CRAN.R-project.org/package=SDAResources}
}
\usage{
Pckg_SDAResources_gpa_H1
}
\description{
This data set is derived from the data(gpa). The design is 1-stage cluster
sampling with clusters of equal sizes. Each cluster (suite) has 4 elements
with the same weight. The target variable is gpa.
}
\keyword{datasets}
