% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pckg_survey_apiclus2_v2_H1}
\alias{Pckg_survey_apiclus2_v2_H1}
\title{A Clustered RDBESDataObject converted from package survey dataset apiclus2}
\format{
A list containing entries required for H1 RDBES data:
\describe{
\item{DE}{the Design data table. Contains 1 DE row with DEstratumName == "Pckg_SDAResources_apiclus2_v2_H1"}
\item{SD}{the Sampling Details data table. Contains 1 child SD row}
\item{VS}{the Vessel Selection data table. Contains 126 child rows (the 126 schools finally observed), each associated to its cluster (dname), VSnumberTotalClusters is 757, VSnumberTotal is 1...5 schools sampled}
\item{FT}{the Fishing Trip data table. Just 1:1 links to the final data (in SA)}
\item{FO}{the Fishing Operation data table. Just 1:1 links to the final data (in SA)}
\item{SS}{the Species Selection data table. Just 1:1 links to the final data (in SA)}
\item{SA}{the Sample data table. SAsampleWeightMeasured is enroll (note the 4 NAs)}
\item{FM}{the Frequency Measure data table}
\item{BV}{the Biological Variable data table}
\item{VD}{the Vessel Details data table}
\item{SL}{the Species List data table}
\item{IS}{the Individual Species table}
}
}
\source{
\url{https://CRAN.R-project.org/package=survey}
}
\usage{
Pckg_survey_apiclus2_v2_H1
}
\description{
This data set is derived from the Academic Performance Index computed for all
California schools based on standardized testing of students. The original
data sets contain information for all schools with at least 100 students and
for various probability samples of the data. The design is 2-stage cluster
sampling with clusters of unequal sizes. An SRS of 40 districts is selected
(psus) from the 757 districts in the population and then up to 5 schools (min
\enumerate{
\item were selected from each district (ssus). The target variable is enroll - note that it contains 4 NA values.
}
}
\keyword{datasets}
