% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESEstObject.R
\name{prepareSubSampleLevelLookup}
\alias{prepareSubSampleLevelLookup}
\title{Private function to get sub-sample level and top-level SAid for SA data}
\usage{
prepareSubSampleLevelLookup(SAdata)
}
\arguments{
\item{SAdata}{The SA data to check}
}
\value{
A data.table with SAid, topLevelSAid and subSampleLevel
}
\description{
Private function to get sub-sample level and top-level SAid for SA data
}
