% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESEstObject.R
\name{processRDBESEstObjLowerHierarchies}
\alias{processRDBESEstObjLowerHierarchies}
\title{Private function to process the lower hierarchies when creating
the RDBESEstObject}
\usage{
processRDBESEstObjLowerHierarchies(rdbesPrepObject, verbose = FALSE)
}
\arguments{
\item{rdbesPrepObject}{A prepared RDBESRawObj}
}
\value{
allLower - the FM and BV tables combined
}
\description{
Private function to process the lower hierarchies when creating
the RDBESEstObject
}
