% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESEstObject.R
\name{processRDBESEstObjUpperHierarchies}
\alias{processRDBESEstObjUpperHierarchies}
\title{Private function to process the upper hierarchies when creating
the RDBESEstObject}
\usage{
processRDBESEstObjUpperHierarchies(
  myRDBESEstObj = NULL,
  rdbesPrepObject,
  hierarchyToUse,
  i = 1,
  verbose = FALSE
)
}
\arguments{
\item{myRDBESEstObj}{An RDBESEstObj to add data to}

\item{rdbesPrepObject}{A prepared RDBESRawObj}

\item{hierarchyToUse}{The hierarchy we are using}

\item{i}{Interger to keep track of where we are in the list of tables}
}
\value{

}
\description{
Private function to process the upper hierarchies when creating
the RDBESEstObject
}
