% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESEstObject.R
\name{procRDBESEstObjLowHier}
\alias{procRDBESEstObjLowHier}
\title{Private function to process the lower hierarchies when creating
the RDBESEstObject}
\usage{
procRDBESEstObjLowHier(rdbesPrepObject, verbose = FALSE)
}
\arguments{
\item{rdbesPrepObject}{A prepared RDBESRawObj}

\item{verbose}{logical. Output messages to console.}
}
\value{
allLower - the FM and BV tables combined
}
\description{
Private function to process the lower hierarchies when creating
the RDBESEstObject
}
