% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESEstObject.R
\name{procRDBESEstObjUppHier}
\alias{procRDBESEstObjUppHier}
\title{Private function to process the upper hierarchies when creating
the RDBESEstObject}
\usage{
procRDBESEstObjUppHier(
  myRDBESEstObj = NULL,
  rdbesPrepObject,
  hierarchyToUse,
  i = 1,
  targetTables,
  verbose = FALSE
)
}
\arguments{
\item{myRDBESEstObj}{An RDBESEstObj to add data to}

\item{rdbesPrepObject}{A prepared RDBESRawObj}

\item{hierarchyToUse}{The hierarchy we are using}

\item{i}{Integer to keep track of where we are in the list of tables}

\item{targetTables}{The RDBES tables we are interested in}
}
\value{

}
\description{
Private function to process the upper hierarchies when creating
the RDBESEstObject
}
