% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRDBESDataObject.R,
%   R/sortRDBESDataObject.R, R/summaryRDBESDataObject.R
\name{print.RDBESDataObject}
\alias{print.RDBESDataObject}
\alias{sort.RDBESDataObject}
\alias{summary.RDBESDataObject}
\title{Print method for RDBESDataObject}
\usage{
\method{print}{RDBESDataObject}(x, ...)

\method{sort}{RDBESDataObject}(x, decreasing = TRUE, ...)

\method{summary}{RDBESDataObject}(object, ...)
}
\arguments{
\item{x}{An object of class RDBESDataObject.}

\item{...}{parameters to underling functions (not used currently)}

\item{decreasing}{should hierarchy tables be the first ones}

\item{object}{An object of class RDBESDataObject.}
}
\value{
None.

The sorted RDBESDataObject by hierarchy.

A list with three elements:
\itemize{
\item{hierarchy: The hierarchy of the DE data.table in the RDBESDataObject.}
\item{rows: A named list where the names are the names of the data.tables in the RDBESDataObject and
the values are the number of rows in each data.table. NULL values are excluded.}
\item{CS: A logical value indicating if the hierarchy is not NULL.}
}
}
\description{
This method prints the hierarchy of the DE data.table (if it exists),
and the number of rows for each data.table in the RDBESDataObject that is not NULL.
It also provides the sampling method and number sampled and number total for tables
where it is applicable.
If the RDBESDataObject has a mixed hierarchy, a warning message is printed.

This method sorts the RDBESDataObject based on the hierarchy.

This method returns a list containing the hierarchy of the DE data.table,
the number of rows for each data.table in the RDBESDataObject that is not NULL,
and a logical value indicating if the hierarchy is not NULL.
}
\examples{
# Print the package data object
print(H1Example)
# Sort the package data
sort(H8ExampleEE1)
# Get summary of the package data
summary(H1Example)
}
