% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePrefixFromVarNames.R
\name{removePrefixFromVarNames}
\alias{removePrefixFromVarNames}
\title{Remove table prefix from variable names}
\usage{
removePrefixFromVarNames(x)
}
\arguments{
\item{x}{RDBES raw object}
}
\value{
updated RDBES raw object where table prefix has been removed from
all variables names except ids
}
\description{
Remove table prefix from variable names
}
\examples{
\dontrun{
myH1RawObject <-
  importRDBESDataCSV(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19")
cleanPrefixFromVarNames(x = myH1RawObject)
}
}
