% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runChecksOnSelectionAndProbs.R
\name{runChecksOnSelectionAndProbs}
\alias{runChecksOnSelectionAndProbs}
\title{Run basic checks on selection methods and probabilities}
\usage{
runChecksOnSelectionAndProbs(x, verbose = FALSE, strict = TRUE)
}
\arguments{
\item{x}{- RDBES raw object}

\item{verbose}{- If TRUE prints the issue behind the stop}

\item{strict}{(Optional) This function validates its input data - should
the validation be strict? The default is TRUE.}
}
\value{
nothing
}
\description{
This function runs some basic checks on selection methods and
and probabilities of the different sampling tables of a hierarchy.
It should be run ahead of generateProbs to secure its correct execution
and for that reason it is included in the wrapper applyGenerateProbs.
}
\seealso{
\code{\link{applyGenerateProbs}}
\code{\link{generateProbs}}

examples for now see
https://github.com/ices-eg/WK_RDBES/tree/master/WKRDB-EST2/chairs/Nuno
}
