% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRDBESDataObjectDataTypes.R
\name{setRDBESDataObjectDataTypes}
\alias{setRDBESDataObjectDataTypes}
\title{For a given RDBESDataObject convert the required
columns to the correct data types.  (This function can cause an error if we
have data in the columns that can't be cast to the desired data type.)}
\usage{
setRDBESDataObjectDataTypes(RDBESDataObjectToConvert)
}
\arguments{
\item{RDBESDataObjectToConvert}{list - the raw item for conversion}
}
\value{
An RDBESDataObject with the correct date types for the required
fields
}
\description{
For a given RDBESDataObject convert the required
columns to the correct data types.  (This function can cause an error if we
have data in the columns that can't be cast to the desired data type.)
}
