% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRDBESRawObjectDataTypes.R
\name{setRDBESRawObjectDataTypes}
\alias{setRDBESRawObjectDataTypes}
\title{For a given rdbesRawObject convert the required
columns to the correct data types.  (This function can cause an error if we
have data in the columns that can't be cast to the desired data type.)}
\usage{
setRDBESRawObjectDataTypes(rdbesRawObjectToConvert)
}
\arguments{
\item{rdbesRawObjectToConvert}{list - the raw item for conversion}
}
\value{
An rdbesRawObject with the correct date types for the required
fields
}
\description{
For a given rdbesRawObject convert the required
columns to the correct data types.  (This function can cause an error if we
have data in the columns that can't be cast to the desired data type.)
}
