% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tablesInRDBESHierarchies}
\alias{tablesInRDBESHierarchies}
\title{The tables required for each RDBES hierarchy.}
\format{
A data frame containing the tables required for each RDBES hierachy.
\describe{
  \item{hierarchy}{the hierachy this applies to H1 to H13}
  \item{table}{the 2-letter table name}
  \item{lowerHierarchy}{is this a lower hierarchy table?}
  \item{optional}{is this table optional within the hierarchy?}
  \item{samplingUnit}{is this table a sampling unit within the hierarchy?}
  \item{sortOrder}{the table sort order within the hiaerarchy}
}
}
\source{
\url{https://github.com/davidcurrie2001/MI_RDBES_ExchangeFiles}
}
\usage{
tablesInRDBESHierarchies
}
\description{
A data frame containing the tables required for each RDBES hierachy
}
\keyword{datasets}
