% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateRDBESDataObject.R
\name{validateRDBESDataObject}
\alias{validateRDBESDataObject}
\alias{checkRDBESDataObject}
\title{Check whether an RDBESDataObject is in a valid format}
\usage{
validateRDBESDataObject(objectToCheck, checkDataTypes = FALSE, verbose = FALSE)

checkRDBESDataObject(objectToCheck, checkDataTypes = FALSE, verbose = FALSE)
}
\arguments{
\item{objectToCheck}{RDBESDataObject i.e. a list of data.tables}

\item{checkDataTypes}{(Optional) Set to TRUE if you want to check that
the data types of the required columns are correct, or FALSE if you don't
care. Default value is FALSE.}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}
}
\value{
TRUE if object is valid, FALSE is object is not valid
}
\description{
Check whether an RDBESDataObject is in a valid format
}
\examples{
\dontrun{
myH1RawObject <-
createRDBESDataObject(rdbesExtractPath = "tests/testthat/h1_v_1_19")
validateRDBESDataObject(myH1RawObject)}

}
