% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateRDBESDataObjectDuplicates.R
\name{validateRDBESDataObjectDuplicates}
\alias{validateRDBESDataObjectDuplicates}
\title{check RDBES Raw Object Content
 Private function to do some basic checks on
the content of the RDBESDataObject (e.g. all required field names are
present).  Function is only used by checkRDBESDataObject and should only
be passed a list of non-null objects}
\usage{
validateRDBESDataObjectDuplicates(
  objectToCheck,
  verbose = FALSE,
  strict = TRUE
)
}
\arguments{
\item{objectToCheck}{- RDBESDataObject i.e. a list of data.tables}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}

\item{strict}{(Optional) Set to TRUE if you want to be sure all columns
are present in the data, set to FALSE if you only want to check that
essential columns are present.  The default is TRUE.}
}
\value{
list with  first element as the object and the second the warnings
}
\description{
check RDBES Raw Object Content
 Private function to do some basic checks on
the content of the RDBESDataObject (e.g. all required field names are
present).  Function is only used by checkRDBESDataObject and should only
be passed a list of non-null objects
}
