% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateRDBESDataObjectFieldNames.R
\name{validateRDBESDataObjectFieldNames}
\alias{validateRDBESDataObjectFieldNames}
\title{check RDBES Data Object field names
Private function to do some checks on the columns of an RDBESDataObject -
1) are all required fields present? 2) are there any extra fields present?
It is used by validateRDBESDataObject() and should only be passed a list
of non-null objects}
\usage{
validateRDBESDataObjectFieldNames(
  objectToCheck,
  verbose = FALSE,
  strict = TRUE
)
}
\arguments{
\item{objectToCheck}{- RDBESDataObject i.e. a list of data.tables}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}

\item{strict}{(Optional) Set to TRUE if you want to be sure all columns
are present in the data, set to FALSE if you only want to check that
essential columns are present.  The default is TRUE.}
}
\value{
list with  first element as a boolean indicating validity
and the second element contains any warnings
}
\description{
check RDBES Data Object field names
Private function to do some checks on the columns of an RDBESDataObject -
1) are all required fields present? 2) are there any extra fields present?
It is used by validateRDBESDataObject() and should only be passed a list
of non-null objects
}
