% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wormsAphiaRecord}
\alias{wormsAphiaRecord}
\title{A dataset containing aphia records for species found in icesSpecWoRMS}
\format{
A data frame
\describe{
\item{AphiaID}{E.g. 100684 }
\item{url}{E.g. "https://www.marinespecies.org/aphia.php?p=taxdetails&id=100684"}
\item{scientificname}{E.g. "Cerianthidae"  }
\item{authority}{E.g. "Milne Edwards & Haime, 1851" }
\item{status}{E.g. "accepted"  }
\item{unacceptreason}{E.g. NA }
\item{taxonRankID}{E.g. 140  }
\item{rank}{E.g. "Family" "Genus" "Species" "Species" }
\item{valid_AphiaID}{E.g. 100684 }
\item{valid_name}{E.g. "Cerianthidae"  }
\item{valid_authority}{E.g. "Milne Edwards & Haime, 1851" }
\item{parentNameUsageID}{E.g. 151646 }
\item{kingdom}{E.g. "Animalia"  }
\item{phylum}{E.g. "Cnidaria"  }
\item{class}{E.g. "Anthozoa"  }
\item{order}{E.g. "Spirularia" }
\item{family}{E.g. "Cerianthidae"  }
\item{genus}{E.g. NA "Cerianthus"}
\item{citation }{E.g. "Molodtsova, T. (2023). World List of Ceriantharia.
Cerianthidae Milne Edwards & Haime, 1851. Accessed through: "... }
\item{isMarine}{E.g. 1 }
\item{isBrackish}{E.g. 1 }
\item{isFreshwater}{E.g. 0  }
\item{isTerrestrial}{E.g. 0  }
\item{isExtinct}{E.g. NA  }
\item{match_type}{E.g. "exact" }
\item{modified}{E.g. "2018-01-22T17:48:34.063Z" }
\item{DateDownloaded}{E.g. "2023-10-18" }
  ...
}
}
\source{
\url{https://www.marinespecies.org/}
}
\usage{
wormsAphiaRecord
}
\description{
A dataset containing aphia records for species found in icesSpecWoRMS
}
\keyword{datasets}
