% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineRDBESDataObjects.R
\name{combineRDBESDataObjects}
\alias{combineRDBESDataObjects}
\title{Combine Two RDBES Raw Objects
combines 2 RDBESDataObjects into a single RDBESDataObject by merging individual
tables one by one}
\usage{
combineRDBESDataObjects(RDBESDataObject1, RDBESDataObject2, strict = TRUE)
}
\arguments{
\item{RDBESDataObject1}{The first object to combine}

\item{RDBESDataObject2}{The second object to combine}

\item{strict}{(Optional) This function validates its input data - should
the validation be strict? The default is TRUE.}
}
\value{
the combination of \code{RDBESDataObject1} and  \code{RDBESDataObject2}
}
\description{
Combine Two RDBES Raw Objects
combines 2 RDBESDataObjects into a single RDBESDataObject by merging individual
tables one by one
}
\examples{
\dontrun{

myH1RawObject <-
    importRDBESDataCSV(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19")
myH5RawObject <-
    importRDBESDataCSV(rdbesExtractPath = "tests\\\\testthat\\\\h5_v_1_19")

myCombinedRawObject <- combineRDBESDataObjects(RDBESDataObject1=myH1RawObject,
                                             RDBESDataObject2=myH5RawObject)
}
}
\seealso{
\link[data.table]{rbindlist}
}
