% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESDataObject.R
\name{createRDBESDataObject}
\alias{createRDBESDataObject}
\title{Create an RDBES Data Object}
\usage{
createRDBESDataObject(
  x = NULL,
  listOfFileNames = NA,
  castToCorrectDataTypes = TRUE
)
}
\arguments{
\item{x}{The path to the zip file extracted from RDBES, or path to a folder
of \code{csv} files, or a list object in the current environment containing data frames of each table.}

\item{listOfFileNames}{(Optional) A named list of file names - the list names
shoudl be the two-letter code for the relevent table e.g. list("DE" =
"DE.csv",... ).  If the parameter is not supplied then the default file
names used by the RDBES data download will be used e.g. "Design.csv" etc.}

\item{castToCorrectDataTypes}{(Optional) If TRUE then the function will
attempt to cast the required columns to the correct data type.  If FALSE
then the column data types will be determined by how the csv files are read
in.  The default is TRUE}
}
\value{
A RDBESDataObject.  If a path to RDBES extract files is provided then
it will contain the data from those files.  If no path is supplied then an
empty RDBESDataObject will be returned.
}
\description{
Create a RDBES Data object from either:
\itemize{
\item a \code{zip} file downloaded from RDBES (or multiple zip files if you want to include CL and CE data)
\item a folder containing \code{csv} files downloaded from RDBES (e.g. the unzipped archive file)
\item a \code{list} of data frames in the current environment representing different tables in the hierarchy
}
}
\examples{
myEmptyRDBESObject <- importRDBESDataCSV()
}
