% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESEstObject.R
\name{createRDBESEstObject}
\alias{createRDBESEstObject}
\title{Creates an RDBESEstObject from RDBES data}
\usage{
createRDBESEstObject(
  rdbesPrepObject,
  hierarchyToUse = NULL,
  stopTable = NULL,
  verbose = FALSE,
  strict = TRUE,
  incDesignVariables = TRUE
)
}
\arguments{
\item{rdbesPrepObject}{The RDBES object that should be used to
create an estimation object}

\item{hierarchyToUse}{The upper RDBES hierarchy to use}

\item{stopTable}{(Optional) The table to stop at in the RDBES hierarchy.
If specified, only tables up to and including this table will be included in the
resulting RDBESEstObject. The default is NULL, which means all tables in the hierarchy
will be included.}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}

\item{strict}{(Optional) This function validates its input data - should
the validation be strict? The default is TRUE.}

\item{incDesignVariables}{(Optional) Should the design variables be included?
The default is TRUE.}
}
\value{
An object of class RDBESEstObject ready for use in design based
estimation
}
\description{
Creates an RDBESEstObject from RDBES data
}
\examples{
myH1EstObj <- createRDBESEstObject(H1Example, 1, "SA")
}
