% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRDBESEstObject.R
\name{createRDBESEstObject}
\alias{createRDBESEstObject}
\title{Creates an rdbesEStObject from prepared RDBES data}
\usage{
createRDBESEstObject(
  rdbesPrepObject,
  hierarchyToUse,
  stopTable = NULL,
  verbose = FALSE
)
}
\arguments{
\item{rdbesPrepObject}{The prepared RDBES object that should be used to
create an estimation object}

\item{hierarchyToUse}{The upper RDBES hiearchy to use}

\item{stopTable}{PLEASE DOCUMENT THIS}

\item{verbose}{PLEASE DOCUMENT THIS}
}
\value{
An object of class RDBESEstObject ready for us in design based
estimation
}
\description{
Creates an rdbesEStObject from prepared RDBES data
}
\examples{
\dontrun{
myH1RawObject <-
  createRDBESDataObject(rdbesExtractPath = "tests/testthat/h1_v_1_19")
myH1EstObj <- createRDBESEstObject(myH1RawObject, 1)
}
}
