% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEstimationForAllStrata.R
\name{doEstimationForAllStrata}
\alias{doEstimationForAllStrata}
\title{Estimate totals and means, and try to generate samples variances for all
strata in an RDBESRawObject}
\usage{
doEstimationForAllStrata(
  rdbesRawObjectForEstim,
  hierarchyToUse,
  verbose = TRUE
)
}
\arguments{
\item{rdbesRawObjectForEstim}{The RDBESRawObject to generate estimates for}

\item{hierarchyToUse}{The number of the RDBES hierarchy to estimate for}

\item{verbose}{(Optional) If set to TRUE more detailed text will be printed
out by the function.  Default is TRUE.}
}
\value{
A data frame containing estimates for all strata
}
\description{
Estimate totals and means, and try to generate samples variances for all
strata in an RDBESRawObject
}
\examples{
\dontrun{

myH1RawObject <-
  createRDBESRawObject(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19")

# Update our test data with some random sample measurements
myH1RawObject[["SA"]]$SAsampWtLive <-
  round(runif(n = nrow(myH1RawObject[["SA"]]), min = 1, max = 100))

myStrataEst <- doEstimationForAllStrata(
  rdbesRawObjectForEstim = myH1RawObject,
  hierarchyToUse = 1
)
}
}
