% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRDBESDataZip.R
\name{downloadRDBESDataZip}
\alias{downloadRDBESDataZip}
\title{Title}
\usage{
downloadRDBESDataZip(
  datatype = c("CE", "CL", "VD", "SL"),
  year,
  hierarchy = NULL,
  country = "FR",
  export_format = "TableWithIdsFormat",
  dir = ".",
  verbose = TRUE
)
}
\arguments{
\item{datatype}{description}

\item{year}{description}

\item{hierarchy}{description}

\item{country}{description}

\item{export_format}{description}

\item{dir}{description}

\item{verbose}{description}
}
\value{
result
}
\description{
need checks on argument format
}
\details{
it seems that datatype = "CS" require a hierarchy value
}
