% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAndTidyRDBESDataObject.R
\name{filterAndTidyRDBESDataObject}
\alias{filterAndTidyRDBESDataObject}
\title{Filter and remove orphan records in an RDBESDataObject}
\usage{
filterAndTidyRDBESDataObject(
  RDBESDataObjectToFilterAndTidy,
  fieldsToFilter,
  valuesToFilter,
  killOrphans = TRUE,
  verbose = FALSE,
  strict = TRUE
)
}
\arguments{
\item{RDBESDataObjectToFilterAndTidy}{The RDBESDataObject to filter.}

\item{fieldsToFilter}{A vector of the field names you wish to check.}

\item{valuesToFilter}{A vector of the field values you wish to filter for.}

\item{killOrphans}{Controls if orphan rows are removed. Default is TRUE.}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}

\item{strict}{(Optional) This function validates its input data - should
the validation be strict? The default is TRUE.}
}
\value{
The filtered input object of the same class as \code{RDBESDataObjectToFilterAndTidy}.
}
\description{
This function filters an RDBESDataObject based on specified fields and values, and can optionally remove any orphan records.
The returned object will include all rows which either: a) do not include any of the field names in \code{fieldsToFilter}, or b) do include the field names and have one of the allowed values in \code{valuesToFilter}.
If \code{killOrphans} is set to \code{TRUE}, the function will remove orphaned rows. The default is \code{TRUE}.
}
\examples{
\dontrun{

myH1RawObject <- createRDBESDataObject(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19_13")

# To check how removeBrokenVesselLinks() works
myH1RawObject$VD$VDlenCat[which(myH1RawObject$VD$VDencrVessCode=="VDcode_10")] <- "VL40XX"

myFields <- c("VSencrVessCode", "VDlenCat")
myValues <- c("VDcode_1","VDcode_2", "VDcode_10","VL1518","VL2440")

myFilteredObject <- filterAndTidyRDBESDataObject(myH1RawObject,
  fieldsToFilter = myFields,
  valuesToFilter = myValues,
  killOrphans = TRUE, verboseBrokenVesselLinks = TRUE
)
}
}
