% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndKillOrphans.R
\name{findAndKillOrphans}
\alias{findAndKillOrphans}
\title{This function finds and removed any orphan records in an
RDBESRawObject.  Normally data that has been downloaded from the RDBES
will not contain orphan records - however if the data is subsequently
filtered it is possible to introduce orphan records.}
\usage{
findAndKillOrphans(objectToCheck, verbose = TRUE)
}
\arguments{
\item{objectToCheck}{an RDBESRawObject.}

\item{verbose}{(Optional) If set to TRUE more detailed text will be printed
out by the function.  Default is TRUE.}
}
\value{
an RDBESRawObject with any orphan records removed
}
\description{
This function finds and removed any orphan records in an
RDBESRawObject.  Normally data that has been downloaded from the RDBES
will not contain orphan records - however if the data is subsequently
filtered it is possible to introduce orphan records.
}
\examples{
\dontrun{

myH1RawObject <-
createRDBESRawObject(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19")
myFields <- c("SDctry","VDctry","VDflgCtry","FTarvLoc")
myValues <- c("ZW","ZWBZH","ZWVFA" )
myFilteredObject <- filterRDBESRawObject(myH1RawObject,
                                        fieldsToFilter = myFields,
                                        valuesToFilter = myValues )
myObjectNoOrphans <- findAndKillOrphans(objectToCheck = myFilteredObject,
                                       verbose = FALSE)
}
}
