% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMissingSSRows.R
\name{generateMissingSSRows}
\alias{generateMissingSSRows}
\title{Generate any missing SS rows.
When FOcatchReg=="All" it is expected that SScatchFraction is either "Catch"
OR "Lan"+"Dis". In the latter case, if one is missing the other is to
be assumed 0.
This function generates SS rows for any missing catch fractions.}
\usage{
generateMissingSSRows(
  RDBESDataObject,
  speciesListName,
  verbose = FALSE,
  strict = TRUE
)
}
\arguments{
\item{RDBESDataObject}{A valid RDBESDataObject}

\item{speciesListName}{The name of the Species List you want to use
for any SS rows that are created.}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}

\item{strict}{(Optional) This function validates its input data - should
the validation be strict? The default is TRUE.}
}
\value{
A data table of SS data with any missing rows added
}
\description{
Generate any missing SS rows.
When FOcatchReg=="All" it is expected that SScatchFraction is either "Catch"
OR "Lan"+"Dis". In the latter case, if one is missing the other is to
be assumed 0.
This function generates SS rows for any missing catch fractions.
}
\examples{
# To follow
}
