% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateZerosUsingSL.R
\name{generateZerosUsingSL}
\alias{generateZerosUsingSL}
\title{Generate zeros in samples using Species List information}
\usage{
generateZerosUsingSL(x)
}
\arguments{
\item{x}{RDBES data frame}
}
\value{
RDBES data frame where SA was complemented with species looked for
(sensu in sampling objectives) but not registered in sample
}
\description{
Generate zeros in samples using Species List information
}
