% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doEstimationForAllStrata.R,
%   R/doEstimationForAllStrataWithEstObject.R
\name{getEstimForStratum}
\alias{getEstimForStratum}
\title{Private function used by doEstimationForAllStrata to get the estimates}
\usage{
getEstimForStratum(x)

getEstimForStratum(x)
}
\arguments{
\item{x}{The input}
}
\value{
Whoever revises this function please specify what it returns here

Whoever revises this function please specify what it returns here
}
\description{
Private function used by doEstimationForAllStrata to get the estimates

Private function used by doEstimationForAllStrata to get the estimates
}
