% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H7Example}
\alias{H7Example}
\title{A dataset containing test RDBES data for H7 in the RDBESDataObject structure}
\format{
A list containing entries required for H7 RDBES data:
\describe{
\item{DE}{the Design data table}
\item{SD}{the Sampling Details data table}
\item{OS}{the Onshore Sample data table}
\item{LE}{the Landing Event data table}
\item{SS}{the Species Selection data table}
\item{SA}{the Sample data table}
\item{BV}{the Biological Variable data table}
\item{SL}{the Species List data table}
\item{IS}{the Individual Species table}
}
#' @source Richard Meitern @ Estonian Marine Institute, 2025
}
\usage{
H7Example
}
\description{
This dataset does not have passed the RDBES upload checks,
hence the object might be somewhat invalid, however it resembles real data
from the Estonian Market Sampling for 2022 for 2 species
}
\keyword{datasets}
