% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H8ExampleEE1}
\alias{H8ExampleEE1}
\title{A dataset containing test RDBES data for H8 in the RDBESDataObject structure}
\format{
A list containing entries required for H8 RDBES data:
\describe{
\item{DE}{the Design data table}
\item{SD}{the Sampling Details data table}
\item{TE}{the Temporal Event data table}
\item{VS}{the Vessel Selection data table}
\item{LE}{the Landing Event data table}
\item{SS}{the Species Selection data table}
\item{SA}{the Sample data table}
\item{BV}{the Biological Variable data table}
\item{VD}{the Vessel Details data table}
\item{SL}{the Species List data table}
\item{CL}{the Commertial Landing data table}
\item{CE}{the Commertial Effort data table}
}
#' @source Richard Meitern @ Estonian Marine Institute, 2023
}
\usage{
H8ExampleEE1
}
\description{
This dataset does not have passed the RDBES upload checks,
hence the object might be somewhat invalid, however it resembles real data
from the Estonian Baltic Trawling fleet for 2022 sprat total landings and
commercial sampling
}
\keyword{datasets}
