% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icesSpecWoRMS}
\alias{icesSpecWoRMS}
\title{ICES Species (WoRMS) code list snapshot}
\format{
A data frame with the following columns:
\describe{
\item{CodeTypeGUID}{GUID of the code type in ICES Vocabulary (e.g. the 'Species (WoRMS)' list).}
\item{CodeTypeID}{Numeric ID of the code type.}
\item{Guid}{GUID identifying this code record.}
\item{Id}{Numeric ID of this code record.}
\item{Key}{AphiaID (numeric key from WoRMS).}
\item{Description}{Scientific name.}
\item{LongDescription}{(If present) additional description; often not used.}
\item{Modified}{Datetime the record was last modified at ICES.}
\item{Deprecated}{Logical; whether this code is deprecated at ICES.}
\item{DateDownloaded}{Date the snapshot was downloaded, e.g. "2023-10-18".}
}
}
\source{
\url{https://vocab.ices.dk/}
}
\usage{
icesSpecWoRMS
}
\description{
A dataset containing a copy of the ICES 'Species (WoRMS)' code list.
The latest code list can be downloaded from https://vocab.ices.dk/.
}
\keyword{datasets}
