% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRDBESDataZIP.R
\name{importRDBESDataZIP}
\alias{importRDBESDataZIP}
\title{Import RDBES Downloaded Data}
\usage{
importRDBESDataZIP(
  filenames,
  castToCorrectDataTypes = TRUE,
  Hierarchy = NULL,
  ...
)
}
\arguments{
\item{filenames}{\itemize{
\item path to zipfile. Multiple paths can be entered if, for
instance, you want to add CL and CE tables to the same object.
}}

\item{castToCorrectDataTypes}{Logical. If \code{TRUE} then the function will
attempt to cast the required columns to the correct data type.  If \code{FALSE}
then the column data types will be determined by how the csv files are read
in. Default is \code{TRUE}.
@param Hierarchy - Optional. A number specifying the hierarchy of the data
to be imported. If the .zip file contains multiple hierarchies.}
}
\value{
a list of all the RDBES data tables The table that are not in input
data are NULL
}
\description{
An internal function used by \code{createRDBESDataObject}. Reads the .zip files
downloadable from the ICES RDBES web page. The function accepts a filepaths
to zip files, unzips and import tables as an object of class
\code{RDBESDataObject}.
}
\examples{
\dontrun{
rdbesExtractPath <- "./2022_FPN_FPE_H7.zip"
obj <- importRDBESDataZIP(rdbesExtractPath)
}
}
\keyword{internal}
