% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRDBESDownloadData.R
\name{importRDBESDataZIP}
\alias{importRDBESDataZIP}
\title{Import RDBES Downloaded Data}
\usage{
importRDBESDataZIP(filenames, castToCorrectDataTypes = TRUE)
}
\arguments{
\item{filenames}{- vector of paths pointing to files that should be imported}

\item{castToCorrectDataTypes}{(Optional) If TRUE then the function
will attempt to cast the required columns to the correct data type.  If
FALSE then the column data types will be determined by how the csv files
are read in.  The default is TRUE.}
}
\value{
a list of all the RDBES data tables
The table that are not in input data are NULL
}
\description{
Read the .zip files and/or .csv files downloadable from the ICES RDBES
web page. The function accepts a list of paths to csv and zip files. Unzips
an then uses \code{\link{importRDBESDataCSV}}
}
\examples{
files <- c("./tests/testthat/h1_v_1_19/H1_2021_000_example.zip")
obj <- importRDBESDataZIP(files)
}
\seealso{
\code{\link{importRDBESDataCSV}}
}
