% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBrokenSpeciesListLinks.R
\name{removeBrokenSpeciesListLinks}
\alias{removeBrokenSpeciesListLinks}
\title{Remove rows which are not pointing to a valid SpecliestListDetails (SL)
records i.e.those rows which have a value of SpeciesListName that does not
exist in the SL table.}
\usage{
removeBrokenSpeciesListLinks(objectToCheck, verbose = FALSE)
}
\arguments{
\item{objectToCheck}{an RDBESDataObject.}

\item{verbose}{(Optional) If set to TRUE more detailed text will be printed
out by the function.  Default is TRUE.}
}
\value{
an RDBESDataObject with any records with an invalid SpeciesListName
rows removed
}
\description{
Remove rows which are not pointing to a valid SpecliestListDetails (SL)
records i.e.those rows which have a value of SpeciesListName that does not
exist in the SL table.
}
\examples{
\dontrun{

myH1RawObject <-
  createRDBESDataObject(rdbesExtractPath = "tests\\\\testthat\\\\h1_v_1_19")
myFields <- c("SLspeclistName")
myValues <- c("WGRDBES-EST TEST 5 - sprat data")
myFilteredObject <- filterRDBESDataObject(myH1RawObject,
  fieldsToFilter = myFields,
  valuesToFilter = myValues
)
myObjectValidSpeciesListLinks <- removeBrokenSpeciesListLinks(
  objectToCheck = myFilteredObject,
  verbose = FALSE
)
}
}
