% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSAwithTaxonFromSL.R
\name{updateSAwithTaxonFromSL}
\alias{updateSAwithTaxonFromSL}
\title{Function which changes the value of SAspeCode in SA.}
\usage{
updateSAwithTaxonFromSL(
  RDBESDataObject,
  validate = TRUE,
  verbose = FALSE,
  strict = TRUE
)
}
\arguments{
\item{RDBESDataObject}{An RDBESDataObject.}

\item{validate}{Set to TRUE if you want validation to be carried out. The
default if TRUE.}

\item{verbose}{(Optional) Set to TRUE if you want informative text on validation printed
out, or FALSE if you don't.  The default is FALSE.}

\item{strict}{(Optional) This function can validate its input data - should
the validation be strict? The default is TRUE.}
}
\value{
RDBES data object where species in SA were renaming for species
occuring in SL for level of species rank. If in SA is
Sprat(126425), in SL Clupeidae (125464) function renameSpeciesSA rename Sprat
from SA to Clupeidae. Clupeidae(family rank) is higher rank than Sprat(species rank).
}
\description{
Function checks the rank of aphia id in both of tables SA and SL, and tries
to replace a more accurate rank in SA with a broader rank from SL.
There are 3 possible situations:
\enumerate{
\item If the aphiaid id in the SA table is more accurate than in the SL table
and the aphiaids are from the same kingdom, phylum, class, order, family,
or genus then the aphiaid in the SA table is changed to the aphia id from
the SL table.
\item If the aphia Ids are from different kingdom, phylum, class, order, family,
or genus then the function retains the original SA species code.
\item If the SL table has a more accurate rank than in the SA table, the
function also retains the original SA species code.
}
}
\examples{
\dontrun{
myObject <- createRDBESDataObject(input = "WGRDBES-EST/personal/Kasia/vignettes/vignetteData")
renameSpeciesSA(RDBESDataObject=myObject,validate,verbose,strict)}
}
