% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateRDBESEstObject.R
\name{validateRDBESEstObject}
\alias{validateRDBESEstObject}
\title{Check whether an object is a valid RDBESEstObject}
\usage{
validateRDBESEstObject(objectToCheck, verbose = FALSE)
}
\arguments{
\item{objectToCheck}{The object to check}

\item{verbose}{(Optional) Set to TRUE if you want informative text printed
out, or FALSE if you don't.  The default is FALSE.}
}
\value{
Whoever revises this function please specify what it returns here
}
\description{
Check whether an object is a valid RDBESEstObject
}
\examples{
\dontrun{
myH1RawObject <-
importRDBESDataCSV(rdbesExtractPath = "tests/testthat/h1_v_1_19")
myEStObj <- createRDBESEstObject(myH1RawObject,1)
validateRDBESEstObject(myEStObj)}
}
